requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
requests
python-dotenv
werkzeug
markdown-it-py
fastapi
python-multipart
uvicorn
markdown
sse-starlette
uuid
faiss-cpu
matplotlib
sentence-transformers
langchain
langchain-openai
langchain-community
langchain-ollama
flask-cors

